fix(ndi): RTCP-liveness signal = peer RR fingerprint, not transport bytes (#388)
Functional verification on dev2 (gst 1.24.2) revealed the original
transport-bytes reader was wrong for this GStreamer version:
- the `transport` get-stats struct carries NO `bytes-received` on gst 1.24,
  so the reader returned None every pass and reaped LIVE consumers after the
  stale window (false positive — verified: a kept-alive Chrome consumer was
  wrongly reaped at t+70s);
- the nested `octets-received`/`packets-received` counters that DO advance live
  on the internal SENDER source — but they track what WE send, so they advance
  for a zombie too (we keep encoding to nobody), which would reap NO ONE.

The only stats driven exclusively by the PEER are its RTCP receiver reports
(`remote-inbound-rtp` round-trip-time/jitter + the `rb-*` RR block in
`gst-rtpsource-stats`). peer_rr_fingerprint hashes those volatile fields: it
CHANGES every sample while the peer reports and freezes (or the struct
vanishes → 0 sentinel) when the peer is gone. liveness_is_stale now treats a
CHANGED fingerprint (not a `>` byte count) as progress.

Verified live on dev2:
- live Chrome consumer kept alive 95s → never reaped (fingerprint changes);
- page killed without WHEP DELETE → reaped within one stale window (~66-71s),
  reason="rtcp-stale" with a frozen fingerprint logged;
- normal connect → WHEP DELETE → slot freed immediately.

Also split the zombie-reaper subsystem out of consumers.rs into a new
pipeline/reaper.rs module — the additions pushed consumers.rs over the
1000-line file hard cap. reaper.rs holds the reap methods + liveness helpers;
consumers.rs keeps build/negotiation/snapshot. #388
